+++ src/Platform.h
@@ -2,6 +2,7 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
@@ -62,4 +63,4 @@
 int initializePlatform(void);
 void cleanupPlatform(void);
 
-unsigned long long PltGetMillis(void);
+uint64_t PltGetMillis(void);
